requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Mail==0.10.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pillow==11.0.0
pycryptodome==3.21.0
Werkzeug==3.1.3
